pyproject.toml
[project]
name = "reconraven"
version = "0.1.0-alpha"
description = "RTL-SDR based SIGINT platform for signal intelligence and RF analysis"
authors = [
    {name = "ReconRaven Team"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "pyrtlsdr>=0.2.93",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "flask>=3.0.0",
    "flask-socketio>=5.3.0",
    "flask-cors>=4.0.0",
    "python-socketio>=5.10.0",
    "python-engineio>=4.8.0",
    "matplotlib>=3.7.0",
    "Pillow>=10.0.0",
    "gpsd-py3==0.3.0",
    "pyserial==3.5",
    "pyyaml==6.0.1",
    "psutil>=5.9.0",
    "pyusb>=1.2.0",
    "pyargus>=1.1.0",
    "soundfile>=0.12.0",
    "openai-whisper>=20231117",
    "colorlog>=6.7.0",
    "python-dateutil>=2.8.0",
    "typing-extensions>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.4",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.7.0",
]

[project.scripts]
reconraven = "reconraven.cli:main"

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
# Target Python 3.9+
target-version = "py39"

# Line length
line-length = 100

# Exclude directories
exclude = [
    ".git",
    ".ruff_cache",
    "__pycache__",
    "build",
    "dist",
    ".venv",
    "venv",
    "recordings",
    "frequency_cache",
]

[tool.ruff.lint]
# Enable pycodestyle (E, W), Pyflakes (F), isort (I), pydocstyle (D)
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "ICN",    # flake8-import-conventions
    "PIE",    # flake8-pie
    "T20",    # flake8-print (catch print statements)
    "PYI",    # flake8-pyi
    "PT",     # flake8-pytest-style
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented code)
    "PL",     # pylint
    "TRY",    # tryceratops
    "RUF",    # ruff-specific rules
]

# Disable specific rules that conflict with project needs
ignore = [
    "E501",    # line too long (we have line-length = 100)
    "D100",    # missing docstring in public module (will add incrementally)
    "D101",    # missing docstring in public class
    "D102",    # missing docstring in public method
    "D103",    # missing docstring in public function
    "D104",    # missing docstring in public package
    "D107",    # missing docstring in __init__
    "T201",    # print found (we'll fix these)
    "ISC001",  # conflicts with formatter
    "Q000",    # conflicts with formatter (single quotes)
    "Q003",    # conflicts with formatter
    "TRY003",  # long messages in exceptions
    "PLR0913", # too many arguments
    "PLR2004", # magic value comparison
    "PTH123",  # pathlib open (legacy code uses open())
]

[tool.ruff.lint.per-file-ignores]
# Tests can use assert statements and magic values
"tests/**/*.py" = ["S101", "PLR2004"]
# CLI can use print statements for user output
"reconraven.py" = ["T201"]
"reconraven/cli.py" = ["T201"]

[tool.ruff.lint.isort]
known-first-party = ["reconraven"]
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
# Use single quotes
quote-style = "single"
indent-style = "space"
line-ending = "auto"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # Enable gradually
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--verbose",
    "--color=yes",
]

